feat: 토큰 발급 함수 경로 변경 (#31)
- 경로: `libs/common/authentication` -> `libs/utils`
- 이름: TokenService -> SignToken
