+++ server/apps/api/src/users/users.module.ts
@@ -2,11 +2,12 @@ import { Module } from '@nestjs/common';
 import { UsersController } from './users.controller';
 import { UsersService } from './users.service';
 import { UsersRepository } from './users.repository';
-import { TokenService } from 'libs/common/authentication/token.service';
 import { JwtService } from '@nestjs/jwt';
+import { SignToken } from 'libs/utils/sign-token';
+import { ConfigService } from '@nestjs/config';
 
 @Module({
   controllers: [UsersController],
-  providers: [UsersService, UsersRepository, TokenService, JwtService],
+  providers: [UsersService, UsersRepository, ConfigService, JwtService, SignToken],
 })
 export class UsersModule {}
