+++ server/libs/utils/sign-token/index.ts
@@ -1,0 +1,1 @@
+export * from './sign-token';
