+++ server/libs/utils/sign-token/token-option.ts
@@ -1,0 +1,14 @@
+export const tokenOption = {
+  accessToken: (jwtSecret) => {
+    return {
+      expiresIn: '1hr',
+      secret: jwtSecret,
+    };
+  },
+  refreshToken: (jwtSecret) => {
+    return {
+      expiresIn: '14d',
+      secret: jwtSecret,
+    };
+  },
+};
